+++ lib/openapi_first/utils.rb
@@ -11,5 +11,21 @@ module OpenapiFirst
     def self.classify(string)
       Hanami::Utils::String.classify(string)
     end
+
+    def self.deep_stringify(params) # rubocop:disable Metrics/MethodLength
+      params.each_with_object({}) do |(key, value), output|
+        output[key.to_s] =
+          case value
+          when ::Hash
+            deep_stringify(value)
+          when Array
+            value.map do |item|
+              item.is_a?(::Hash) ? deep_stringify(item) : item
+            end
+          else
+            value
+          end
+      end
+    end
   end
 end